♻️ refactor: community.tsx をディレクトリ構成に分割 + ルート遅延読み込み
- community.tsx (468行) → community/ (3ファイル: community-page, community-header, index)
- main.tsx: BenchmarkPage, WorldMapPage を React.lazy 化
- non-null assertion を optional chain に置換
