+++ apps/web/src/pages/community/community-page.tsx
@@ -1,4 +1,4 @@
-import { AppProvider, Canvas, TransientLayer, useApp } from "@edv4h/usketch-canvas-engine";
+import { AppProvider, Canvas, TransientLayer } from "@edv4h/usketch-canvas-engine";
 import { type AppInstance, createApp } from "@edv4h/usketch-core";
 import { createDomRendererPlugin } from "@edv4h/usketch-dom-renderer";
 import { createActivityFeedPlugin } from "@edv4h/usketch-plugin-activity-feed";
@@ -29,9 +29,12 @@ import { createBoardStore } from "@edv4h/usketch-store";
 import { createWsProvider, type WsProviderHandle } from "@edv4h/usketch-sync";
 import { useEffect, useRef, useState } from "react";
 import { useNavigate, useParams } from "react-router";
-import { api } from "../lib/api.js";
-import { getDevUser } from "../lib/dev-auth.js";
-import { useAuth } from "../lib/use-auth.js";
+import { api } from "../../lib/api.js";
+import { getDevUser } from "../../lib/dev-auth.js";
+import { getErrorMessage } from "../../lib/errors.js";
+import { useAuth } from "../../lib/use-auth.js";
+import { useKeyboardShortcuts } from "../../lib/use-keyboard-shortcuts.js";
+import { CommunityHeader } from "./community-header.js";
 
 export function CommunityPage() {
 	const { slug } = useParams<{ slug: string }>();
@@ -81,7 +84,7 @@ export function CommunityPage() {
 							if (isPublic) {
 								await api.boards.update(board.id, { isPublic: true });
 							}
-							store!.updateShape(shapeId, {
+							store?.updateShape(shapeId, {
 								boardId: board.id,
 								boardTitle: board.title,
 								ownerName: authUserName ?? "",
@@ -90,7 +93,7 @@ export function CommunityPage() {
 							});
 						} catch (e) {
 							console.error("Failed to create board:", e);
-							store!.deleteShape(shapeId);
+							store?.deleteShape(shapeId);
 						}
 					},
 				});
@@ -104,7 +107,7 @@ export function CommunityPage() {
 							wsUrl += `?devUserId=${encodeURIComponent(devUser.id)}`;
 						}
 					}
-					wsProvider = createWsProvider({ url: wsUrl, doc: syncHandle!.doc });
+					wsProvider = createWsProvider({ url: wsUrl, doc: syncHandle?.doc });
 					wsProviderRef.current = wsProvider;
 
 					extraPlugins.push(createRippleEffectPlugin(wsProvider));
@@ -199,11 +202,11 @@ export function CommunityPage() {
 					createDomRendererPlugin(),
 				];
 
-				await syncHandle!.whenSynced;
-				if (cancelled) return;
+				await syncHandle?.whenSynced;
+				if (cancelled || !store) return;
 
 				const plugins = [...basePlugins, ...extraPlugins];
-				const created = await createApp({ store: store!, plugins });
+				const created = await createApp({ store, plugins });
 				if (cancelled) {
 					created.destroy();
 					return;
@@ -236,7 +239,7 @@ export function CommunityPage() {
 				setApp(instance);
 			} catch (e) {
 				if (!cancelled) {
-					setError(e instanceof Error ? e.message : "Failed to initialize community");
+					setError(getErrorMessage(e, "Failed to initialize community"));
 				}
 			}
 		})();
@@ -252,44 +255,7 @@ export function CommunityPage() {
 	}, [slug, authUserId, authUserName, authUserImage, navigate]);
 
 	// キーボードショートカット
-	useEffect(() => {
-		if (!app) return;
-
-		const handleKeyDown = (e: KeyboardEvent) => {
-			const tag = (e.target as HTMLElement)?.tagName;
-			const isInput =
-				tag === "INPUT" || tag === "TEXTAREA" || (e.target as HTMLElement)?.isContentEditable;
-
-			// Escape はテキスト入力中でも通す（入力欄からblurしてツールをselectに戻す）
-			if (e.key === "Escape" && isInput) {
-				(e.target as HTMLElement)?.blur();
-				app.shortcuts.handleKeyDown(e);
-				return;
-			}
-
-			// テキスト入力中はそれ以外のショートカットを無視
-			if (isInput) {
-				return;
-			}
-			const tools = app.tools.getAll();
-			for (const [id, def] of tools) {
-				if (
-					def.shortcut &&
-					e.key.toLowerCase() === def.shortcut.toLowerCase() &&
-					!e.ctrlKey &&
-					!e.metaKey &&
-					!e.altKey
-				) {
-					app.store.setActiveToolId(id);
-					return;
-				}
-			}
-			app.shortcuts.handleKeyDown(e);
-		};
-
-		window.addEventListener("keydown", handleKeyDown);
-		return () => window.removeEventListener("keydown", handleKeyDown);
-	}, [app]);
+	useKeyboardShortcuts(app);
 
 	if (error) {
 		return (
@@ -309,160 +275,5 @@ export function CommunityPage() {
 				<CommunityHeader regionName={regionName} />
 			</div>
 		</AppProvider>
-	);
-}
-
-function CommunityHeader({ regionName }: { regionName: string }) {
-	const navigate = useNavigate();
-	const { user: sessionUser, logout } = useAuth();
-	const app = useApp();
-
-	return (
-		<>
-			<div
-				style={{
-					position: "fixed",
-					top: 12,
-					left: 12,
-					zIndex: 100,
-					display: "flex",
-					gap: 8,
-					alignItems: "center",
-				}}
-			>
-				<button
-					type="button"
-					onClick={() => navigate("/community")}
-					style={{
-						background: "white",
-						border: "none",
-						borderRadius: 8,
-						padding: "6px 12px",
-						boxShadow: "0 2px 8px rgba(0,0,0,0.12)",
-						fontSize: 14,
-						cursor: "pointer",
-						color: "#0066ff",
-						fontFamily: "system-ui, sans-serif",
-					}}
-				>
-					World Map
-				</button>
-				<div
-					style={{
-						background: "white",
-						borderRadius: 8,
-						padding: "6px 14px",
-						boxShadow: "0 2px 8px rgba(0,0,0,0.12)",
-						fontSize: 14,
-						fontWeight: 600,
-						fontFamily: "system-ui, sans-serif",
-					}}
-				>
-					{regionName || "uSketch"}
-				</div>
-				{sessionUser ? (
-					<button
-						type="button"
-						onClick={() => {
-							logout();
-							navigate("/login");
-						}}
-						style={{
-							background: "white",
-							border: "none",
-							borderRadius: 8,
-							padding: "6px 12px",
-							boxShadow: "0 2px 8px rgba(0,0,0,0.12)",
-							fontSize: 12,
-							cursor: "pointer",
-							color: "#666",
-						}}
-					>
-						{sessionUser.name} — Sign Out
-					</button>
-				) : (
-					<a
-						href="/login"
-						style={{
-							background: "white",
-							borderRadius: 8,
-							padding: "6px 12px",
-							boxShadow: "0 2px 8px rgba(0,0,0,0.12)",
-							fontSize: 12,
-							textDecoration: "none",
-							color: "#0066ff",
-						}}
-					>
-						Sign In
-					</a>
-				)}
-			</div>
-			{/* 右上: サイドパネル開閉ボタン */}
-			{sessionUser && (
-				<div
-					style={{
-						position: "fixed",
-						top: 12,
-						right: 12,
-						zIndex: 100,
-						display: "flex",
-						gap: 6,
-					}}
-				>
-					<button
-						type="button"
-						onClick={() => app.events.emit("side-panel:toggle", { tabId: "board-info" })}
-						style={{
-							background: "white",
-							border: "none",
-							borderRadius: 8,
-							padding: "6px 10px",
-							boxShadow: "0 2px 8px rgba(0,0,0,0.12)",
-							fontSize: 14,
-							cursor: "pointer",
-							color: "#475569",
-						}}
-						title="Board Info"
-					>
-						📋
-					</button>
-					<button
-						type="button"
-						onClick={() => app.events.emit("side-panel:toggle", { tabId: "comments" })}
-						style={{
-							background: "white",
-							border: "none",
-							borderRadius: 8,
-							padding: "6px 10px",
-							boxShadow: "0 2px 8px rgba(0,0,0,0.12)",
-							fontSize: 14,
-							cursor: "pointer",
-							color: "#475569",
-						}}
-						title="Comments"
-					>
-						💬
-					</button>
-					<button
-						type="button"
-						onClick={() => app.events.emit("side-panel:toggle", { tabId: "community-chat" })}
-						style={{
-							background: "white",
-							border: "none",
-							borderRadius: 8,
-							padding: "6px 10px",
-							boxShadow: "0 2px 8px rgba(0,0,0,0.12)",
-							fontSize: 14,
-							cursor: "pointer",
-							color: "#475569",
-						}}
-						title="Chat"
-						aria-label="Chat"
-					>
-						🗨️
-					</button>
-				</div>
-			)}
-		</>
 	);
 }
