+++ apps/web/src/pages/community/index.ts
@@ -1,0 +1,1 @@
+export { CommunityPage } from "./community-page.js";
